Rename bounded-while-loop tutorial
